collector/requirements.txt
requests
pandas
yfinance

# Supabase stack
supabase
gotrue
postgrest
storage3
realtime

# Mongo
pymongo
python-dotenv
